+++ 2key-protocol/test/unitTests/envRelatedTests/twoKeyExchangeRate.spec.ts
@@ -1,9 +1,10 @@
-import {exchangeRates} from "../../constants/smallConstants";
 import {expect} from 'chai';
 import 'mocha';
 import web3Switcher from "../../helpers/web3Switcher";
 import {TwoKeyProtocol} from "../../../src";
 import getTwoKeyProtocol from "../../helpers/twoKeyProtocol";
+import {promisify} from "../../../src/utils/promisify";
+import singletons from "../../../src/contracts/singletons";
 
 require('es6-promise').polyfill();
 require('isomorphic-fetch');
@@ -17,6 +18,30 @@ const usdDaiSymbol = 'USD-DAI';
 const usd2KeySymbol = '2KEY-USD';
 const tusdSymbol = 'TUSD-USD';
 const daiSymbol = 'DAI-USD';
+const daiEthSymbol = 'DAI-ETH'
+const tusdEthSymbol = 'TUSD-ETH'
+
+let oraclesObjects = [];
+
+let oracles = [
+    usdSymbol,
+    usdDaiSymbol,
+    usd2KeySymbol,
+    tusdSymbol,
+    daiSymbol,
+    daiEthSymbol,
+    tusdEthSymbol
+];
+
+let exchangeRates = {
+    'USD': 100,
+    'USD-DAI': 1.03,
+    '2KEY-USD': 0.06,
+    'TUSD-USD': 0.99,
+    'DAI-USD': 0.97,
+    'DAI-ETH': 0.0097,
+    'TUSD-ETH': 0.0099
+}
 
 
 describe(
@@ -36,46 +61,88 @@ describe(
       }
     );
 
-    it('should set dollar rate', async () => {
-      const usdRate = 30;
-      const txHash = await twoKeyProtocol.TwoKeyExchangeContract.setValue(
-        usdSymbol,
-        usdRate,
-        from
-      );
-      await twoKeyProtocol.Utils.getTransactionReceiptMined(txHash);
-      const value = await twoKeyProtocol.TwoKeyExchangeContract.getBaseToTargetRate(usdSymbol);
+    it('should deploy oracles using manager contract', async() => {
 
-      expect(value).to.be.eq(usdRate);
+        let managerAddress = await twoKeyProtocol.SingletonRegistry.getNonUpgradableContractAddress('MockOraclesManager');
+        let descriptions = [];
+
+        for(const oracle of oracles) {
+            descriptions.push(twoKeyProtocol.Utils.toHex(oracle));
+        }
+
+        // Create manager instance
+        let managerInstance = twoKeyProtocol.web3.eth.contract(singletons.MockOraclesManager.abi).at(managerAddress);
+
+        let txHash = await promisify(managerInstance.deployAndStoreOracles,[
+            18,
+            descriptions,
+            1,
+            {from: from}
+        ]);
+
+        // At this certain point we can assume that all oracles are freshly deployed
+        await twoKeyProtocol.Utils.getTransactionReceiptMined(txHash);
+
+        let [oracleAddresses,deployedDescriptions] = await promisify(managerInstance.getDeployedOraclesAndDescriptions,[]);
+
+        console.log(oracleAddresses, deployedDescriptions);
+        let counter = 0;
+
+        for (const description of deployedDescriptions) {
+            let oracleAddressFromContract = await promisify(managerInstance.pairToOracleAddress,[description]);
+            expect(oracleAddressFromContract.toString().toLowerCase())
+                .to.be.equal(oracleAddresses[counter].toString().toLowerCase());
+
+            counter++;
+        }
+
+        // Now it's time to store this oracles inside TwoKeyExchangeRateContract
+        let txHash1 = await promisify(twoKeyProtocol.twoKeyExchangeContract.storeChainLinkOracleAddresses,[
+            descriptions,
+            oracleAddresses,
+            {from: from}
+        ]);
+
     }).timeout(timeout);
 
 
-    it('should set dollar/dai rate', async () => {
-      const usdDaiRate = 50;
+    it('should set rates on oracles', async() => {
+        let rates = [];
+        let oracleNamesHex = [];
 
-      const txHash = await twoKeyProtocol.TwoKeyExchangeContract.setValue(
-        usdDaiSymbol,
-        usdDaiRate,
-        from);
-      await twoKeyProtocol.Utils.getTransactionReceiptMined(txHash);
-      const value = await twoKeyProtocol.TwoKeyExchangeContract.getBaseToTargetRate(usdDaiSymbol);
+        let managerAddress = await twoKeyProtocol.SingletonRegistry.getNonUpgradableContractAddress('MockOraclesManager');
 
-      expect(value).to.be.eq(usdDaiRate);
+        // Create manager instance
+        let managerInstance = twoKeyProtocol.web3.eth.contract(singletons.MockOraclesManager.abi).at(managerAddress);
+
+        for (const [key, value] of Object.entries(exchangeRates)) {
+            rates.push(twoKeyProtocol.Utils.toWei(value));
+        }
+
+        for(const oracle of oracles) {
+            oracleNamesHex.push(twoKeyProtocol.Utils.toHex(oracle));
+        }
+
+
+        let txHash = await promisify(managerInstance.updateRates,[
+            oracleNamesHex,
+            rates,
+            {
+                from
+            }
+        ]);
+
+        let receipt = await twoKeyProtocol.Utils.getTransactionReceiptMined(txHash);
     }).timeout(timeout);
 
-    it('should set dollar and dollar/dai rates with one request', async () => {
-      const usdRate = exchangeRates.usd;
-      const usdDaiRate = exchangeRates.usdDai;
-      const usd2KeyRate = exchangeRates.usd2Key;
-      const daiUsdRate = exchangeRates.dai;
-      const tusdUsdRate = exchangeRates.tusd;
 
+    it('should check dollar dai, dollar 2Key rates', async () => {
+      const usdRate = exchangeRates.USD;
+      const usdDaiRate = exchangeRates["USD-DAI"];
+      const usd2KeyRate = exchangeRates["2KEY-USD"];
+      const daiUsdRate = exchangeRates["DAI-USD"];
+      const tusdUsdRate = exchangeRates["TUSD-USD"];
 
-      const txHash = await twoKeyProtocol.TwoKeyExchangeContract.setValues(
-        [usdSymbol, usdDaiSymbol, usd2KeySymbol, daiSymbol, tusdSymbol],
-        [usdRate, usdDaiRate, usd2KeyRate, daiUsdRate, tusdUsdRate],
-        from);
-      await twoKeyProtocol.Utils.getTransactionReceiptMined(txHash);
       let usdValue = await twoKeyProtocol.TwoKeyExchangeContract.getBaseToTargetRate(usdSymbol);
       let usdDaiValue = await twoKeyProtocol.TwoKeyExchangeContract.getBaseToTargetRate(usdDaiSymbol);
       let usd2KeyValue = await twoKeyProtocol.TwoKeyExchangeContract.getBaseToTargetRate(usd2KeySymbol);
@@ -85,7 +152,7 @@ describe(
     }).timeout(timeout);
 
     it('should correctly exchange eth to dollar', async () => {
-      const usdRate = exchangeRates.usd;
+      const usdRate = exchangeRates.USD;
       const weiAmount = 100;
 
       const setRateTxHash = await twoKeyProtocol.TwoKeyExchangeContract.setValue(
@@ -103,20 +170,17 @@ describe(
     }).timeout(timeout);
 
     it('should get fiat to stable coin quotes', async () => {
-
         let amountFiatWei = parseFloat(twoKeyProtocol.Utils.toWei(15,'ether').toString());
         let pairs = await twoKeyProtocol.TwoKeyExchangeContract.getFiatToStableQuotes(amountFiatWei, 'USD', ['DAI','TUSD']);
         console.log(pairs);
-
     }).timeout(timeout);
 
     it('should get stable coin quota by address', async () => {
       let daiAddress = await twoKeyProtocol.SingletonRegistry.getNonUpgradableContractAddress('DAI');
-      console.log(daiAddress);
       let quota = await twoKeyProtocol.TwoKeyExchangeContract.getStableCoinToUSDQuota(
         daiAddress
       );
-      console.log(quota);
+      expect(quota).to.be.equal(exchangeRates["DAI-USD"]);
     })
   }
 );