+++ scripts/deployments/test-public/DONE-13.01.2020-29e908c3836566576987cc5715ebadfbc0f626f2.bash
@@ -17,6 +17,8 @@ echo "Destination for execution of contracts patches: 0xf4797416e6b6835114390591
 
 cd ../..
 
-python3 generate_bytecode.py upgradeContract TwoKeyUpgradableExchange 1.0.37
+python3 generate_bytecode.py upgradeContract TwoKeyUpgradableExchange 1.0.40
+python3 generate_bytecode.py upgradeContract TwoKeyExchangeRateContract 1.0.4
+python3 generate_bytecode.py approveNewCampaign TOKEN_SELL 1.0.46