+++ truffle.js
@@ -133,7 +133,7 @@ module.exports = {
           skipDryRun: true,
           network_id: ids["prod-public"],
           gas: 8000000,
-          gasPrice: 25000000000,
+          gasPrice: 69000000000,
       },
 
       'public.prod-hdwallet': {