+++ CHANGELOG.md
@@ -2,6 +2,78 @@
 ## [Unreleased]
 
 
+<a name="v1.4.190-develop"></a>
+## [v1.4.190-develop] - 2021-01-20
+
+<a name="v1.4.189-develop"></a>
+## [v1.4.189-develop] - 2021-01-20
+### Reverts
+- Contracts deployed to public.staging-hdwallet Jan 20, 2021 2:48 PM
+
+
+<a name="v1.4.36-master"></a>
+## [v1.4.36-master] - 2021-01-19
+
+<a name="v1.2.77-staging"></a>
+## [v1.2.77-staging] - 2021-01-15
+
+<a name="v1.4.35-master"></a>
+## [v1.4.35-master] - 2021-01-14
+
+<a name="v1.4.34-master"></a>
+## [v1.4.34-master] - 2021-01-14
+
+<a name="v1.4.188-develop"></a>
+## [v1.4.188-develop] - 2021-01-13
+
+<a name="v1.2.76-staging"></a>
+## [v1.2.76-staging] - 2021-01-13
+
+<a name="v1.2.75-staging"></a>
+## [v1.2.75-staging] - 2021-01-12
+
+<a name="v1.2.74-staging"></a>
+## [v1.2.74-staging] - 2021-01-12
+
+<a name="v1.4.187-develop"></a>
+## [v1.4.187-develop] - 2021-01-12
+
+<a name="v1.4.186-develop"></a>
+## [v1.4.186-develop] - 2021-01-12
+
+<a name="v1.4.185-develop"></a>
+## [v1.4.185-develop] - 2021-01-12
+
+<a name="v1.4.184-develop"></a>
+## [v1.4.184-develop] - 2021-01-12
+### Pull Requests
+- Merge pull request [#60](https://github.com/2key/contracts/issues/60) from 2key/oracles
+
+
+<a name="v1.2.73-staging"></a>
+## [v1.2.73-staging] - 2020-12-08
+
+<a name="v1.4.33-master"></a>
+## [v1.4.33-master] - 2020-12-06
+
+<a name="v1.2.72-staging"></a>
+## [v1.2.72-staging] - 2020-12-06
+
+<a name="v1.4.32-master"></a>
+## [v1.4.32-master] - 2020-12-05
+
+<a name="v1.4.31-master"></a>
+## [v1.4.31-master] - 2020-12-05
+
+<a name="v1.4.30-master"></a>
+## [v1.4.30-master] - 2020-12-05
+
+<a name="v1.2.71-staging"></a>
+## [v1.2.71-staging] - 2020-12-02
+
+<a name="v1.2.70-staging"></a>
+## [v1.2.70-staging] - 2020-12-01
+
 <a name="v1.4.183-develop"></a>
 ## [v1.4.183-develop] - 2020-11-30
 
@@ -2246,26 +2318,26 @@
 <a name="development-20180822121004"></a>
 ## [development-20180822121004] - 2018-08-22
 
-<a name="rinkeby-infura-20180822105641"></a>
-## [rinkeby-infura-20180822105641] - 2018-08-22
-
 <a name="rinkeby-infura-20180822105344"></a>
 ## [rinkeby-infura-20180822105344] - 2018-08-22
+
+<a name="rinkeby-infura-20180822105641"></a>
+## [rinkeby-infura-20180822105641] - 2018-08-22
 
 <a name="rinkeby-infura-20180822105105"></a>
 ## [rinkeby-infura-20180822105105] - 2018-08-22
 
-<a name="rinkeby-infura/plasma-ap-20180822095110"></a>
-## [rinkeby-infura/plasma-ap-20180822095110] - 2018-08-22
-
 <a name="dev-local-20180822095208"></a>
 ## [dev-local-20180822095208] - 2018-08-22
 
-<a name="dev-ap/plasma-ap-20180821173644"></a>
-## [dev-ap/plasma-ap-20180821173644] - 2018-08-21
+<a name="rinkeby-infura/plasma-ap-20180822095110"></a>
+## [rinkeby-infura/plasma-ap-20180822095110] - 2018-08-22
 
 <a name="dev-ap/plasma-ap-20180821173916"></a>
 ## [dev-ap/plasma-ap-20180821173916] - 2018-08-21
+
+<a name="dev-ap/plasma-ap-20180821173644"></a>
+## [dev-ap/plasma-ap-20180821173644] - 2018-08-21
 
 <a name="rinkeby-infura/plasma-ap-20180821163715"></a>
 ## [rinkeby-infura/plasma-ap-20180821163715] - 2018-08-21
@@ -2408,7 +2480,29 @@
 <a name="v0.1"></a>
 ## v0.1 - 2018-04-09
 
-[Unreleased]: https://github.com/2key/contracts/compare/v1.4.183-develop...HEAD
+[Unreleased]: https://github.com/2key/contracts/compare/v1.4.190-develop...HEAD
+[v1.4.190-develop]: https://github.com/2key/contracts/compare/v1.4.189-develop...v1.4.190-develop
+[v1.4.189-develop]: https://github.com/2key/contracts/compare/v1.4.36-master...v1.4.189-develop
+[v1.4.36-master]: https://github.com/2key/contracts/compare/v1.2.77-staging...v1.4.36-master
+[v1.2.77-staging]: https://github.com/2key/contracts/compare/v1.4.35-master...v1.2.77-staging
+[v1.4.35-master]: https://github.com/2key/contracts/compare/v1.4.34-master...v1.4.35-master
+[v1.4.34-master]: https://github.com/2key/contracts/compare/v1.4.188-develop...v1.4.34-master
+[v1.4.188-develop]: https://github.com/2key/contracts/compare/v1.2.76-staging...v1.4.188-develop
+[v1.2.76-staging]: https://github.com/2key/contracts/compare/v1.2.75-staging...v1.2.76-staging
+[v1.2.75-staging]: https://github.com/2key/contracts/compare/v1.2.74-staging...v1.2.75-staging
+[v1.2.74-staging]: https://github.com/2key/contracts/compare/v1.4.187-develop...v1.2.74-staging
+[v1.4.187-develop]: https://github.com/2key/contracts/compare/v1.4.186-develop...v1.4.187-develop
+[v1.4.186-develop]: https://github.com/2key/contracts/compare/v1.4.185-develop...v1.4.186-develop
+[v1.4.185-develop]: https://github.com/2key/contracts/compare/v1.4.184-develop...v1.4.185-develop
+[v1.4.184-develop]: https://github.com/2key/contracts/compare/v1.2.73-staging...v1.4.184-develop
+[v1.2.73-staging]: https://github.com/2key/contracts/compare/v1.4.33-master...v1.2.73-staging
+[v1.4.33-master]: https://github.com/2key/contracts/compare/v1.2.72-staging...v1.4.33-master
+[v1.2.72-staging]: https://github.com/2key/contracts/compare/v1.4.32-master...v1.2.72-staging
+[v1.4.32-master]: https://github.com/2key/contracts/compare/v1.4.31-master...v1.4.32-master
+[v1.4.31-master]: https://github.com/2key/contracts/compare/v1.4.30-master...v1.4.31-master
+[v1.4.30-master]: https://github.com/2key/contracts/compare/v1.2.71-staging...v1.4.30-master
+[v1.2.71-staging]: https://github.com/2key/contracts/compare/v1.2.70-staging...v1.2.71-staging
+[v1.2.70-staging]: https://github.com/2key/contracts/compare/v1.4.183-develop...v1.2.70-staging
 [v1.4.183-develop]: https://github.com/2key/contracts/compare/v1.2.69-staging...v1.4.183-develop
 [v1.2.69-staging]: https://github.com/2key/contracts/compare/v1.2.68-staging...v1.2.69-staging
 [v1.2.68-staging]: https://github.com/2key/contracts/compare/v1.4.182-develop...v1.2.68-staging
@@ -3142,14 +3236,14 @@
 [dev-local-20180823081655]: https://github.com/2key/contracts/compare/development-20180822154725...dev-local-20180823081655
 [development-20180822154725]: https://github.com/2key/contracts/compare/development-20180822124421...development-20180822154725
 [development-20180822124421]: https://github.com/2key/contracts/compare/development-20180822121004...development-20180822124421
-[development-20180822121004]: https://github.com/2key/contracts/compare/rinkeby-infura-20180822105641...development-20180822121004
-[rinkeby-infura-20180822105641]: https://github.com/2key/contracts/compare/rinkeby-infura-20180822105344...rinkeby-infura-20180822105641
-[rinkeby-infura-20180822105344]: https://github.com/2key/contracts/compare/rinkeby-infura-20180822105105...rinkeby-infura-20180822105344
-[rinkeby-infura-20180822105105]: https://github.com/2key/contracts/compare/rinkeby-infura/plasma-ap-20180822095110...rinkeby-infura-20180822105105
-[rinkeby-infura/plasma-ap-20180822095110]: https://github.com/2key/contracts/compare/dev-local-20180822095208...rinkeby-infura/plasma-ap-20180822095110
-[dev-local-20180822095208]: https://github.com/2key/contracts/compare/dev-ap/plasma-ap-20180821173644...dev-local-20180822095208
-[dev-ap/plasma-ap-20180821173644]: https://github.com/2key/contracts/compare/dev-ap/plasma-ap-20180821173916...dev-ap/plasma-ap-20180821173644
-[dev-ap/plasma-ap-20180821173916]: https://github.com/2key/contracts/compare/rinkeby-infura/plasma-ap-20180821163715...dev-ap/plasma-ap-20180821173916
+[development-20180822121004]: https://github.com/2key/contracts/compare/rinkeby-infura-20180822105344...development-20180822121004
+[rinkeby-infura-20180822105344]: https://github.com/2key/contracts/compare/rinkeby-infura-20180822105641...rinkeby-infura-20180822105344
+[rinkeby-infura-20180822105641]: https://github.com/2key/contracts/compare/rinkeby-infura-20180822105105...rinkeby-infura-20180822105641
+[rinkeby-infura-20180822105105]: https://github.com/2key/contracts/compare/dev-local-20180822095208...rinkeby-infura-20180822105105
+[dev-local-20180822095208]: https://github.com/2key/contracts/compare/rinkeby-infura/plasma-ap-20180822095110...dev-local-20180822095208
+[rinkeby-infura/plasma-ap-20180822095110]: https://github.com/2key/contracts/compare/dev-ap/plasma-ap-20180821173916...rinkeby-infura/plasma-ap-20180822095110
+[dev-ap/plasma-ap-20180821173916]: https://github.com/2key/contracts/compare/dev-ap/plasma-ap-20180821173644...dev-ap/plasma-ap-20180821173916
+[dev-ap/plasma-ap-20180821173644]: https://github.com/2key/contracts/compare/rinkeby-infura/plasma-ap-20180821163715...dev-ap/plasma-ap-20180821173644
 [rinkeby-infura/plasma-ap-20180821163715]: https://github.com/2key/contracts/compare/rinkeby-infura/plasma-ap-20180821093403...rinkeby-infura/plasma-ap-20180821163715
 [rinkeby-infura/plasma-ap-20180821093403]: https://github.com/2key/contracts/compare/rinkeby-infura/plasma-ap-20180820134542...rinkeby-infura/plasma-ap-20180821093403
 [rinkeby-infura/plasma-ap-20180820134542]: https://github.com/2key/contracts/compare/rinkeby-infura/plasma-ap-20180820094709...rinkeby-infura/plasma-ap-20180820134542